+++ glo-visualizer/src/utils.ts
@@ -1,4 +1,13 @@
 
+export function zip(a, b): any[][] {
+    return Array.from(a)
+        .map((k, i) => [k, Array.from(b)[i]]);
+}
+
+export function isString(value) {
+    return typeof value === 'string' || value instanceof String;
+}
+
 export function getAllChecked(boxes) {
     let boolean = false;
     for (let i = 0; i < boxes.length; i++) {
@@ -96,4 +105,35 @@ export function getStyle(layers, layer) {
     let style = (!layerConfig[0].paint) ? false : layerConfig[0].paint
 
     return style
+}
+
+export function isScrolledIntoView(el) {
+    var rect = el.getBoundingClientRect();
+    var elemTop = rect.top;
+    var elemBottom = rect.bottom;
+    var isVisible = (elemTop >= 0) && (elemBottom <= window.innerHeight);
+    // Partially visible elements return true:
+    //isVisible = elemTop < window.innerHeight && elemBottom >= 0;
+    return isVisible;
+}
+
+export async function fetchJSON (url) {
+    let response = await fetch(url);
+    let data = await response.json();
+    //TODO error handling
+    // console.log(data)
+    return data;
+}
+
+export const fly = async (map,x,y,z,sleep=2000)=> {
+    await new Promise(r => setTimeout(r, sleep));
+    await map.flyTo({
+        center: [x,y],
+        zoom: z,
+        essential: true // this animation is considered essential with respect to prefers-reduced-motion
+        });
+}
+
+export const sleep = async (time_s) => {
+    await new Promise(r => setTimeout(r, time_s*1000))
 }
